backend/requirements.txt
# Core dependencies
python-dotenv==1.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
colorama==0.4.6
python-multipart==0.0.9
fastapi==0.121.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.9
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
websockets==15.0.1

# Authentication
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
python-jose[cryptography]==3.3.0

